Improvements thanks to #1 feedback
Updating this in preparation of YouTube tutorial. Changes thanks to
feedback from @nsthorat. I think if we are to create generic
`NeuralNetwork` or `Classifier` classes in ml5, a lot of the work
will inevitably involve data helper classes so that the end user
can work with vanilla arrays and tensors are all created and managed
internally by ml5.